+++ spec/unit/resources/matchers_spec.rb
@@ -34,7 +34,12 @@ describe 'oneview_test::default' do
     expect(chef_run).to_not create_oneview_ethernet_network('')
     expect(chef_run).to_not bulk_create_oneview_ethernet_network('')
     expect(chef_run).to_not create_oneview_ethernet_network_if_missing('')
+    expect(chef_run).to_not reset_oneview_ethernet_network_connection_template('')
     expect(chef_run).to_not delete_oneview_ethernet_network('')
+
+    # oneview_connection_template
+    expect(chef_run).to_not update_oneview_connection_template('')
+    expect(chef_run).to_not reset_oneview_connection_template('')
 
     # oneview_network_set
     expect(chef_run).to_not create_oneview_network_set('')
@@ -55,6 +60,13 @@ describe 'oneview_test::default' do
     expect(chef_run).to_not update_oneview_logical_enclosure_from_group('')
     expect(chef_run).to_not reconfigure_oneview_logical_enclosure('')
     expect(chef_run).to_not set_oneview_logical_enclosure_script('')
+
+    # oneview_interconnect
+    expect(chef_run).to_not set_oneview_interconnect_uid_light('')
+    expect(chef_run).to_not set_oneview_interconnect_power_state('')
+    expect(chef_run).to_not reset_oneview_interconnect_port_protection('')
+    expect(chef_run).to_not reset_oneview_interconnect('')
+    expect(chef_run).to_not update_oneview_interconnect_port('')
 
     # oneview_logical_interconnect_group
     expect(chef_run).to_not create_oneview_logical_interconnect_group('')