+++ CHANGELOG.md
@@ -2,12 +2,15 @@
   - Upgraded Ruby SDK version to 2.0.0
   - Fixed add/create, delete/remove for resources
   - Added support to Enclosure group actions
-  - Added support for Enclosure `:refresh` and `:reconfigure`
+  - Added support to Enclosure `:refresh` and `:reconfigure`
   - Added support to Ethernet network `:bulk_create`
-  - Added support to Logical Enclosure `:reconfigure`, `:set_script`
+  - Added support to Connection template actions
+  - Integrated Connection template actions in Ethernet network within the update action, and added `:reset_connection_template` action
+  - Added support to Logical enclosure `:reconfigure` and `:set_script`
   - Added support to Network set actions
   - Added support to Datacenter actions
   - Added support to Rack actions
+  - Added support to Interconnect actions
 
 ### 0.1.1
   - Fixed Ruby SDK version to 1.0.0